feat (admin API): added Login repository and service

+++ internal/core/domain/model/user_model.go
@@ -3,10 +3,10 @@ package model
 import "time"
 
 type User struct {
-	ID       uint   `gorm:"id"`
-	Name string `gorm:"name"`
-	Email    string `gorm:"email"`
-	Password string `gorm:"password"`
-	CreatedAt time.Time `gorm:"created_at"`
+	ID        int64      `gorm:"id"`
+	Name      string     `gorm:"name"`
+	Email     string     `gorm:"email"`
+	Password  string     `gorm:"password"`
+	CreatedAt time.Time  `gorm:"created_at"`
 	UpdatedAt *time.Time `gorm:"updated_at"`
 }
